refactoring of files add unit testing
