+++ src/modules/create_graph/pojo/searchNode.py
@@ -1,0 +1,7 @@
+
+class SearchNode:
+
+    def __init__(self, node_id, post_id=None, is_post=False):
+        self.node_id = node_id
+        self.post_id = post_id
+        self.is_post = is_post
